Fixes an issue with running DU sim locally
When DU simulator is run locally, the function crashes due to segfault
inside `nf_oran_du_init` when trying to read `ves_template.json` file.
This is because `ves_template.json` file is not copied during the image
creation.

Also, the actual utility function `file_read_content` instead of
returning a `NULL` crashes, because uninitialized `length` was used to
assign a value when `buffer` was `NULL`. Fixed this as well, the
`buffer[length]` value is only sent if the file exists and a valid
buffer is allocated.

Issue-Id: SIM-109
Change-Id: I75eb985f479331850de2e9fc4aac2ca588fa133c
Signed-off-by: Abhijit Gadgil <gabhijit@iitbombay.org>